Extract Cosmos3 vision scheduler
