+++ tests/unit_test/cosmos3/test_stages.py
@@ -4,13 +4,8 @@ from __future__ import annotations
 
 from types import SimpleNamespace
 
-import torch
-
-from sglang_omni.models.cosmos3 import stages
-from sglang_omni.models.cosmos3.payload_types import Cosmos3PipelineState
-from sglang_omni.proto import OmniRequest, StagePayload
+from sglang_omni.models.cosmos3 import stages, vision_encoder_scheduler
 from sglang_omni.scheduling.simple_scheduler import SimpleScheduler
-from sglang_omni.scheduling.stage_cache import StageOutputCache
 
 
 def test_preprocessing_factory_returns_simple_scheduler(monkeypatch) -> None:
@@ -37,6 +32,37 @@ def test_preprocessing_factory_returns_simple_scheduler(monkeypatch) -> None:
     assert isinstance(scheduler, SimpleScheduler)
     assert scheduler._fn is sentinel
     assert calls == [("nvidia/Cosmos3-Nano", 8192, "cosmos-revision", True)]
+
+
+def test_vision_encoder_factory_delegates_to_scheduler(monkeypatch) -> None:
+    sentinel = object()
+    calls: list[tuple[str, str | None, str, str | None]] = []
+
+    def fake_create(
+        model_path: str,
+        *,
+        revision: str | None = None,
+        device: str,
+        dtype: str | None,
+    ):
+        calls.append((model_path, revision, device, dtype))
+        return sentinel
+
+    monkeypatch.setattr(
+        vision_encoder_scheduler,
+        "create_vision_encoder_scheduler",
+        fake_create,
+    )
+
+    result = stages.create_vision_encoder_executor(
+        "nvidia/Cosmos3-Nano",
+        revision="cosmos-revision",
+        device="cuda:1",
+        dtype="bfloat16",
+    )
+
+    assert result is sentinel
+    assert calls == [("nvidia/Cosmos3-Nano", "cosmos-revision", "cuda:1", "bfloat16")]
 
 
 def test_text_factory_passes_through_tensor_parallelism(
@@ -99,65 +125,3 @@ def test_text_factory_passes_through_tensor_parallelism(
     assert captured["server_args"] is server_args
     assert captured["gpu_id"] == 2
     assert captured["scheduler_kwargs"]["tp_rank"] == 1
-
-
-class _FakeVisionModel:
-    spatial_merge_size = 2
-    out_hidden_size = 4
-    deepstack_layers = 1
-    visual_dtype_bytes = 4
-
-    def __init__(self) -> None:
-        self.calls = 0
-
-    def __call__(self, **inputs):
-        self.calls += 1
-        grid = inputs["image_grid_thw"]
-        token_count = grid.prod(-1) // 4
-        total = int(token_count.sum().item())
-        embeds = torch.arange(total * 4, dtype=torch.float32).reshape(total, 4)
-        return {
-            "image_embeds": embeds,
-            "image_grid_thw": grid,
-            "image_token_counts": token_count,
-            "deepstack_visual_embeds_image": [embeds + 1],
-        }
-
-
-def _vision_payload(request_id: str, cache_key: str) -> StagePayload:
-    state = Cosmos3PipelineState(
-        encoder_inputs={
-            "vision_encoder": {
-                "pixel_values": torch.ones((4, 6)),
-                "image_grid_thw": torch.tensor([[1, 2, 2]]),
-                "cache_key": cache_key,
-            }
-        }
-    )
-    return StagePayload(
-        request_id=request_id,
-        request=OmniRequest(inputs=None),
-        data=state.to_dict(),
-    )
-
-
-def test_vision_encoder_batches_and_deduplicates_same_cache_key() -> None:
-    model = _FakeVisionModel()
-    cache = StageOutputCache(max_size=4, cache_device="cpu")
-    outputs = stages._batch_vision_requests(
-        [
-            _vision_payload("first", "same-image"),
-            _vision_payload("second", "same-image"),
-        ],
-        model=model,
-        cache=cache,
-    )
-
-    assert model.calls == 1
-    assert len(cache) == 1
-    first = Cosmos3PipelineState.from_dict(outputs[0].data)
-    second = Cosmos3PipelineState.from_dict(outputs[1].data)
-    assert torch.equal(
-        first.encoder_outs["vision_encoder"]["image_embeds"],
-        second.encoder_outs["vision_encoder"]["image_embeds"],
-    )
